[UPDATE] Number type이 int로 받아와지는 에러 수정

+++ lib/models/journal/farming_model.dart
@@ -32,7 +32,7 @@ class Farming {
 
   factory Farming.fromMap(Map<String, dynamic> farming) {
     return Farming(
-      farmingArea: farming['farmingArea'],
+      farmingArea: farming['farmingArea'].toDouble(),
       farmingAreaUnit: farming['farmingAreaUnit'],
       farmingMethod: farming['farmingMethod'],
       farmingMethodUnit: farming['farmingMethodUnit'],
